+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:tts/tts.dart';
+import 'dart:math';
 
 void main() => runApp(new MyApp());
 
@@ -10,8 +10,10 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-final TextEditingController controller = new TextEditingController();
-
+final TextEditingController speechTextController = new TextEditingController();
+String languageAvailableText = '';
+List<String> languages;
+Random rng = new Random();
   @override
   initState() {
     super.initState();
@@ -22,16 +24,12 @@ final TextEditingController controller = new TextEditingController();
   initPlatformState() async {
     // Platform messages may fail, so we use a try/catch PlatformException.
     
-
+    languages = await Tts.getAvailableLanguages();
     // If the widget was removed from the tree while the asynchronous platform
     // message was in flight, we want to discard the reply rather than calling
     // setState to update our non-existent appearance.
     if (!mounted)
       return;
-
-    setState(() {
-      
-    });
   }
 
   @override
@@ -45,15 +43,19 @@ final TextEditingController controller = new TextEditingController();
           child: new Column(
             children: <Widget>[
               new TextField(
-                controller: controller,
+                controller: speechTextController,
               ),
               new MaterialButton(
                 color: Colors.blue,
-                  onPressed: (){
-                    Tts.speak(controller.text);
+                  onPressed: () async{
+                    var lang = languages[rng.nextInt(languages.length-1)];
+                    var isGoodLanguage = await Tts.isLanguageAvailable(lang);
+                    var setLanguage = await Tts.setLanguage(lang);
+                    Tts.speak(speechTextController.text);
                   },
                 child: new Text('Do Speak'),
-              )
+              ),
+              new Text(languageAvailableText)
             ],
           ),
         ),