+++ ios/Classes/TtsPlugin.h
@@ -3,4 +3,5 @@
 
 @interface TtsPlugin : NSObject<FlutterPlugin>
 @property (readwrite, nonatomic, strong) AVSpeechSynthesizer *speechSynthesizer;
+@property (strong) NSString *locale;
 @end