+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.0.1] - 2/5/2018.
+
+* Minor bug fix.
+
+## [1.0.0] - 2/5/2018.
+
+* Adding ability to change languages.
+
+## [0.0.2] - 2/4/2018.
+
+* Android bug fix.
+
 ## [0.0.1] - 2/4/2018.
 
-* TODO: initial release with basic speech functionality.
+* initial release with basic speech functionality.